% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Big_Shared_Data.R
\name{Shared_BigData}
\alias{Shared_BigData}
\title{Container Class for data.table Shared Between Tasks}
\description{
Mostly to deal with alloc.col shallow copies, but also nice to have a bit more abstraction.
}
\section{Super class}{
\code{\link[sl3:Shared_Data]{sl3::Shared_Data}} -> \code{Shared_BigData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Shared_BigData$new()}}
\item \href{#method-get_data}{\code{Shared_BigData$get_data()}}
\item \href{#method-add_columns}{\code{Shared_BigData$add_columns()}}
\item \href{#method-clone}{\code{Shared_BigData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shared_BigData$new(data, bm, force_copy = TRUE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shared_BigData$get_data(rows = NULL, columns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_columns"></a>}}
\subsection{Method \code{add_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shared_BigData$add_columns(
  new_data,
  column_uuid = uuid::UUIDgenerate(),
  row_index = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shared_BigData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
